fix ilp64 integer type fault

+++ rstsr-lapack-ffi/scripts/perform_bindgen.py
@@ -129,7 +129,7 @@ use core::ffi::{c_char, c_void};
 #[cfg(not(feature = "ilp64"))]
 pub type blas_int = i32;
 #[cfg(feature = "ilp64")]
-pub type blas_int = i32;
+pub type blas_int = i64;
 """ + "\n\n" + token
 # -
 
@@ -201,7 +201,7 @@ use core::ffi::{c_char, c_void};
 #[cfg(not(feature = "ilp64"))]
 pub type blas_int = i32;
 #[cfg(feature = "ilp64")]
-pub type blas_int = i32;
+pub type blas_int = i64;
 """ + "\n\n" + token
 # -
 
@@ -260,7 +260,7 @@ use core::ffi::c_char;
 #[cfg(not(feature = "ilp64"))]
 pub type lapack_int = i32;
 #[cfg(feature = "ilp64")]
-pub type lapack_int = i32;
+pub type lapack_int = i64;
 """ + "\n\n" + token
 # -
 
@@ -317,7 +317,7 @@ use core::ffi::{c_char, c_int};
 #[cfg(not(feature = "ilp64"))]
 pub type lapack_int = i32;
 #[cfg(feature = "ilp64")]
-pub type lapack_int = i32;
+pub type lapack_int = i64;
 """ + "\n\n" + token
 # -
 
@@ -372,7 +372,7 @@ use core::ffi::{c_char, c_int};
 #[cfg(not(feature = "ilp64"))]
 pub type lapack_int = i32;
 #[cfg(feature = "ilp64")]
-pub type lapack_int = i32;
+pub type lapack_int = i64;
 """ + "\n\n" + token
 # -
