+++ rstsr-openblas-ffi/scripts/perform_bindgen.py
@@ -134,7 +134,7 @@ use core::ffi::{c_char, c_void, c_int};
 #[cfg(not(feature = "ilp64"))]
 pub type blas_int = i32;
 #[cfg(feature = "ilp64")]
-pub type blas_int = i32;
+pub type blas_int = i64;
 
 #[cfg(feature = "quad_precision")]
 #[repr(C)]
@@ -220,7 +220,7 @@ use core::ffi::{c_char, c_void, c_int};
 #[cfg(not(feature = "ilp64"))]
 pub type blas_int = i32;
 #[cfg(feature = "ilp64")]
-pub type blas_int = i32;
+pub type blas_int = i64;
 """ + "\n\n" + token
 # -
