+++ rstsr-openblas-ffi/src/f77blas.rs
@@ -5,7 +5,7 @@ use core::ffi::{c_char, c_int, c_void};
 #[cfg(not(feature = "ilp64"))]
 pub type blas_int = i32;
 #[cfg(feature = "ilp64")]
-pub type blas_int = i32;
+pub type blas_int = i64;
 
 #[cfg(feature = "quad_precision")]
 #[repr(C)]
